Reproducing issue with interfaces within apollo
I've modified the federation demo code here to reproduce an issue we're
seeing in our monolith. Specifically we have the `accounts` service own
an interface (`Task`) and an implementing type (`TestSectionTask`) and
we have the `reviews` service extend both the interface and the
implemementing type.

The issue we're seeing shows up when we make a query like [1] below:

[1]```
{
  me {
    tasksFromReviews {
      reviewsField
      accountsField
    }
  }
}
```

This returns:
```
{
  "data": {
    "me": {
      "tasksFromReviews": [
        null
      ]
    }
  }
}
```

But we expect it to return:
```
{
  "data": {
    "me": {
      "tasksFromReviews": [
        {
          "reviewsField": "reviews",
          "accountsField": "accounts"
        }
      ]
    }
  }
}
```

which _is_ the data we get when we make a query like [2], which fetches
first from `accounts` and then gets data from `reviews`

[2]

```
{
  me {
    tasks {
      reviewsField
      accountsField
    }
  }
}
```

When we run query [1], I expect:
 1. The gateway to fetch a list of Tasks from `reviews`, including the
    reviewsField
 2. The gateway to send a list of representations to the `_entities`
    query to accounts, in order to fetch the `accountsField` for those
    entities
 3. The gateway to merge the data from the accounts with reviews and
    then return to the end user.

But, what we actually see is that step 2 fails - the gateway sends an
empty list of representations to the `accounts` service.

As best as I can tell, the issue is in the query planner (or in the
interaction between query planner and executor). For query [1],
step 1, it outputs a plan like:
```
  Flatten(path: "me") {
      Fetch(service: "reviews") {
        {
          ... on User {
            __typename
            id
          }
        } =>
        {
          ... on User {
            tasksFromReviews {
              __typename
              ... on TestSectionTask {
                reviewsField
                id
              }
            }
          }
        }
      },
    },
  Flatten(path: "me.tasksFromReviews.@") {
      Fetch(service: "accounts") {
        {
          ... on Task {
            __typename
          }
          ... on TestSectionTask {
            id
          }
        } =>
        {
          ... on Task {
            accountsField
          }
        }
      },
    },
```

The issue (i think) is confusingly the lack of `__typename` selected
from the TestSectionTask type - this [3] bit of execute query plan will
only merge in data if the typename matches.

More confusingly, the issue seems to go away when we upgrade to the
latest version of apollo/gateway, but since that version rewrote the
query planner in rust it's not immediately obvious to me what changed
between the JS and rust implementations. We also cannot upgrade to that
version until the issue described in [4] is fixed.

[3] https://github.com/apollographql/federation/blob/88cb9164341363d908c8a414f879ecb143017688/gateway-js/src/executeQueryPlan.ts#L423-L434
[4] https://github.com/apollographql/federation/issues/210